requirements.txt
Flask
PyPDF2
python-docx
fpdf
Pillow